Move PartitionFormat to storage_types.dart
